Crate paralegal_non_rustc_utils

Modules

serde_map_via_vec

Serialize a HashMap by converting it to a Vec, lifting restrictions on the types of permissible keys.

Macros

fs_storable_default

Structs

CommandFactory

DisplayList

Has a Display implementation if the elements of the iterator inside have one. This will render them surrounded by [ brackets and separated by , comma and space

ParalegalArtifact

TruncatedHumanTime

A struct with a Display implementation taht renders a std::time::Duration in human readable form, similar to the humantime crate, but instead of rendering with arbitrary precision it only renders two “significant sections”, e.g. “2h 5min” or “2d 20h”. The sections are days (d), hours (h), minutes (min), seconds (s), miliseconds (ms), microseconds (μs) and nanoseconds (ns).

Constants

ARTIFACT_NAME

FLOW_GRAPH_EXT

FLOW_GRAPH_OUT_NAME

Name of the file used for emitting the serialized [ProgramDescription].

STAT_FILE_EXT

Extension for output files containing statistics of the analzyer run.

Traits

FileSystemStorable

Functions

display_list

Display this iterator as a list

linux_workaround_for_llvm_lib

prepare_analyzer_command

Makes sure paralegal-flow and cargo-paralegal-flow have been built, then returns a factory for Commands that invoke them properly

setup_logging

write_sep

Write all elements from it into the formatter fmt using f, separating them with sep